Simplify setup.py
As this tblib is installed as a Python package, setup.py should use the
'packages' keyword (which it already does) and not 'py_modules'. It is
redundant and unnecessary, so remove it.

This library doesn't have any package data, so don't bother with
'include_package_data'.

This package is just pure Python files, so it is zip safe. Don't bother
setting 'zip_safe' to False.

Remove trove classifiers for older Python implementations that are not
supported.

Remove empty keywords 'install_requires' and 'extras_require'. They are
just noise.